add role assoc to k8s auth method

+++ Lab04/create-k8s-auth.sh
@@ -18,6 +18,7 @@ fi
 
 CODESPACE_DOMAIN="app.github.dev"
 AKEYLESS_GATEWAY_URL="https://${CODESPACE_NAME}-8000.${CODESPACE_DOMAIN}"
+AUTH_METHOD_NAME="/Workshops/Workshop2/${GITHUB_USERNAME}/k8s-auth-method"
 
 echo "Creating ServiceAccount and ClusterRoleBinding..."
 cat << EOF > akl_gw_token_reviewer.yaml
@@ -85,31 +86,38 @@ echo "Extracting K8s Cluster CA Certificate..."
 CA_CERT=$(kubectl config view --raw --minify --flatten \
   --output 'jsonpath={.clusters[].cluster.certificate-authority-data}')
 
-echo "Creating K8s Auth Method..."
-AUTH_RESPONSE=$(akeyless create-auth-method-k8s -n "/Workshops/Workshop2/${GITHUB_USERNAME}/k8s-auth-method" --json)
-ACCESS_ID=$(echo $AUTH_RESPONSE | jq -r .access_id)
-PRV_KEY=$(echo $AUTH_RESPONSE | jq -r .prv_key)
+echo "Checking if K8s Auth Method exists..."
+AUTH_METHOD_EXISTS=$(akeyless auth-method list | jq -r ".auth_methods[] | select(.auth_method_name == \"$AUTH_METHOD_NAME\") | .auth_method_name")
 
-echo "Extracting K8s issuer..."
-K8S_ISSUER=$(kubectl proxy --api-prefix=/k8s-api & \
-  sleep 2 && \
-  curl -s http://localhost:8001/k8s-api/.well-known/openid-configuration | \
-  jq -r .issuer && \
-  kill %1)
+if [ -n "$AUTH_METHOD_EXISTS" ]; then
+    echo "K8s Auth Method already exists. Getting access ID..."
+    ACCESS_ID=$(akeyless auth-method list  | jq -r ".auth_methods[] | select(.auth_method_name == \"$AUTH_METHOD_NAME\") | .auth_method_access_id")
+else
+    echo "Creating K8s Auth Method..."
+    AUTH_RESPONSE=$(akeyless create-auth-method-k8s -n "$AUTH_METHOD_NAME" --json)
+    ACCESS_ID=$(echo $AUTH_RESPONSE | jq -r .access_id)
+    PRV_KEY=$(echo $AUTH_RESPONSE | jq -r .prv_key)
+    echo "Extracting K8s issuer..."
+    K8S_ISSUER=$(kubectl proxy --api-prefix=/k8s-api & \
+      sleep 2 && \
+      curl -s http://localhost:8001/k8s-api/.well-known/openid-configuration | \
+      jq -r .issuer && \
+      kill %1)
 
-# Get cluster endpoint
-K8S_HOST=$(kubectl config view --minify -o jsonpath='{.clusters[0].cluster.server}')
+    # Get cluster endpoint
+    K8S_HOST=$(kubectl config view --minify -o jsonpath='{.clusters[0].cluster.server}')
 
-echo "Creating K8s Gateway Auth Config..."
-akeyless gateway-create-k8s-auth-config \
-  --name "/Workshops/Workshop2/${GITHUB_USERNAME}/k8s-auth-config" \
-  --gateway-url "$AKEYLESS_GATEWAY_URL" \
-  --access-id "$ACCESS_ID" \
-  --signing-key "$PRV_KEY" \
-  --k8s-host "$K8S_HOST" \
-  --token-reviewer-jwt "$SA_JWT_TOKEN" \
-  --k8s-ca-cert "$CA_CERT" \
-  --k8s-issuer "$K8S_ISSUER"
+    echo "Creating K8s Gateway Auth Config..."
+    akeyless gateway-create-k8s-auth-config \
+      --name "$AUTH_METHOD_NAME" \
+      --gateway-url "$AKEYLESS_GATEWAY_URL" \
+      --access-id "$ACCESS_ID" \
+      --signing-key "$PRV_KEY" \
+      --k8s-host "$K8S_HOST" \
+      --token-reviewer-jwt "$SA_JWT_TOKEN" \
+      --k8s-ca-cert "$CA_CERT" \
+      --k8s-issuer "$K8S_ISSUER"
+fi
 
 # Save access ID to a file
 echo "Saving access ID to file..."
@@ -117,4 +125,12 @@ echo "$ACCESS_ID" > k8s_auth_access_id.txt
 
 echo "Auth Method Access ID: $ACCESS_ID"
 echo "Access ID saved to k8s_auth_access_id.txt"
+
+echo "Associating role with auth method..."
+if ! akeyless assoc-role-am --am-name "$AUTH_METHOD_NAME" --role-name "/Workshops/Workshop2" 2>&1 | grep -q "Status 409 Conflict"; then
+    echo "Role association created"
+else
+    echo "Role association already exists"
+fi
+
 echo "Setup complete!"
